fix: add login refresh component under authenticated page
